requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.14
ecdsa==0.18.0
enum-compat==0.0.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
intuit-oauth==1.2.4
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
langchain==0.0.265
langsmith==0.0.22
MarkupSafe==2.1.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==2.4.4
packaging==23.1
pandas==2.0.1
pyasn1==0.5.0
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-quickbooks==0.9.3
pytz==2023.3
PyYAML==6.0.1
rauth==0.7.3
requests==2.30.0
requests-oauthlib==0.8.0
rsa==4.9
simplejson==3.19.1
six==1.16.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
